Refactor environment setup for cross-platform compatibility and streamline path management
